separate tiktok live and mocked tests
since I'll be writing a new test for tiktok, I took the opportunity to
separate the tests. And rename the folder so it is singular as the one
in models.
