+++ test/integration/parser/facebook_item_test.rb
