+++ test/integration/parser/instagram_item_test.rb
