+++ test/integration/parser/kwai_test.rb
